Refactor Handlers to Endpoints and Update README
Refactored `PeopleHandler` to `PeopleEndpoints` and `ProductsHandler` to `ProductsEndpoints`, updating corresponding files and namespaces.
